OCSADV-412: Refactoring: Split some ITC UI code into several files.
