+++ bundle/jsky.app.ot/src/main/scala/jsky/app/ot/editor/seq/ItcPanel.scala
@@ -3,30 +3,20 @@ package jsky.app.ot.editor.seq
 import java.awt.Color
 import javax.swing._
 
-import edu.gemini.itc.shared.PlottingDetails.PlotLimits
 import edu.gemini.itc.shared._
-import edu.gemini.pot.sp.{SPNodeKey, SPComponentType}
-import edu.gemini.spModel.gemini.flamingos2.Flamingos2
-import edu.gemini.spModel.gemini.gmos.GmosNorthType.FPUnitNorth
-import edu.gemini.spModel.gemini.gmos.GmosSouthType.FPUnitSouth
-import edu.gemini.spModel.gemini.gmos.{InstGmosSouth, InstGmosNorth}
-import edu.gemini.spModel.gemini.gsaoi.Gsaoi
-import edu.gemini.spModel.gemini.niri.InstNIRI
-import edu.gemini.spModel.gemini.obscomp.SPSiteQuality.{CloudCover, ImageQuality, SkyBackground, WaterVapor}
+import edu.gemini.pot.sp.SPComponentType
+import jsky.app.ot.itc.{AnalysisMethodPanel, ConditionsPanel, PlotDetailsPanel}
 import jsky.app.ot.util.OtColor
-import jsky.util.gui.{NumberBoxWidget, TextBoxWidget, TextBoxWidgetWatcher}
 import org.jfree.chart.{ChartPanel, JFreeChart}
 
 import scala.concurrent.Future
 import scala.swing.GridBagPanel.{Anchor, Fill}
-import scala.swing.ListView.Renderer
 import scala.swing.ScrollPane.BarPolicy._
+import scala.swing._
 import scala.swing.event._
-import scala.swing.{ButtonGroup, _}
 import scala.util.{Failure, Success}
 
 import scalaz._
-import Scalaz._
 
 object ItcPanel {
 
@@ -268,291 +258,4 @@ private class ItcChartsPanel(table: ItcSpectroscopyTable) extends GridBagPanel {
   }
 
 }
-
-/** User element that allows to change the conditions taken for the calculations on-the-fly. */
-private class ConditionsPanel(owner: EdIteratorFolder) extends GridBagPanel {
-
-  private val ttMsg = "Select conditions for ITC calculations. Values different from program conditions are shown in red."
-
-  class ConditionCB[A](items: Seq[A], renderFunc: A => String) extends ComboBox[A](items) {
-    private var programValue = selection.item
-    tooltip  = ttMsg
-    renderer = new Renderer[A] {
-      override def componentFor(list: ListView[_ <: A], isSelected: Boolean, focused: Boolean, a: A, index: Int): Component = {
-        new Label(renderFunc(a)) {{ foreground = if (programValue == a) Color.BLACK else Color.RED }}
-      }
-    }
-    listenTo(selection)
-    reactions += {
-      case SelectionChanged(_) => foreground = color()
-    }
-
-    def sync(newValue: A) = {
-      if (programValue == selection.item) {
-        // if we are "in sync" with program value (i.e. the program value is currently selected), update it
-        selection.item = newValue
-      }
-      // set new program value and update coloring
-      programValue = newValue
-      foreground = color()
-    }
-
-    private def color()   = if (inSync()) Color.BLACK else Color.RED
-
-    private def inSync()  = programValue == selection.item
-
-  }
-
-  private val sb = new ConditionCB[SkyBackground]  (SkyBackground.values,                       "SB " + _.displayValue())
-  private val cc = new ConditionCB[CloudCover]     (CloudCover.values.filterNot(_.isObsolete),  "CC " + _.displayValue())
-  private val iq = new ConditionCB[ImageQuality]   (ImageQuality.values,                        "IQ " + _.displayValue())
-  private val wv = new ConditionCB[WaterVapor]     (WaterVapor.values,                          "WV " + _.displayValue())
-  private val am = new ConditionCB[Double]         (List(1.0, 1.5, 2.0),                        d => f"Airmass $d%.1f")
-
-  def conditions = new ObservingConditions(
-    iq.selection.item,
-    cc.selection.item,
-    wv.selection.item,
-    sb.selection.item,
-    am.selection.item)
-
-  def update() = {
-    // Note: site quality node can be missing (i.e. null)
-    Option(owner.getContextSiteQuality).foreach { qual =>
-      sb.sync(qual.getSkyBackground)
-      cc.sync(qual.getCloudCover)
-      iq.sync(qual.getImageQuality)
-      wv.sync(qual.getWaterVapor)
-      // TODO: currently the airmass program value is fixed to 1.5 airmass
-      // TODO: can we get this value from the airmass constraints?
-      am.sync(1.5)
-    }
-  }
-
-  tooltip = ttMsg
-
-  layout(new Label("Conditions for ITC calculations:"))  = new Constraints {
-    anchor    = Anchor.West
-    gridx     = 0
-    gridy     = 0
-    gridwidth = 5
-    insets    = new Insets(0, 0, 5, 0)
-  }
-  layout(sb)  = new Constraints {
-    gridx     = 0
-    gridy     = 1
-    insets    = new Insets(0, 0, 0, 0)
-  }
-  layout(cc) = new Constraints {
-    gridx     = 1
-    gridy     = 1
-    insets    = new Insets(0, 3, 0, 0)
-  }
-  layout(iq) = new Constraints {
-    gridx     = 2
-    gridy     = 1
-    insets    = new Insets(0, 3, 0, 0)
-  }
-  layout(wv) = new Constraints {
-    gridx     = 3
-    gridy     = 1
-    insets    = new Insets(0, 3, 0, 0)
-  }
-  layout(am) = new Constraints {
-    gridx     = 4
-    gridy     = 1
-    insets    = new Insets(0, 3, 0, 0)
-  }
-
-  deafTo(this)
-  listenTo(sb.selection, cc.selection, iq.selection, wv.selection, am.selection)
-  reactions += {
-    case SelectionChanged(_) => publish(new SelectionChanged(this))
-  }
-
-}
-
-private object AnalysisMethodPanel {
-  // TODO: The analysis method will have to become part of the observation model and become persistent.
-  // As an intermediate solution (to keep the values available when switching between different
-  // observations for at least as long as the OT is running) we store them here in a local cache.
-  val analysisCache = scala.collection.mutable.Map[SPNodeKey, AnalysisMethod]()
-}
-
-private class AnalysisMethodPanel(owner: EdIteratorFolder) extends GridBagPanel {
-
-  val defaultAnalysisMethod = AutoAperture(5.0)
-
-  val autoAperture  = new RadioButton("Auto") { focusable = false; selected = true }
-  val userAperture  = new RadioButton("User") { focusable = false }
-  val skyLabel      = new Label("Sky Aperture")
-  val skyUnits      = new Label("x target aperture")
-  val sky           = new NumberEdit(skyLabel, skyUnits, 5.0)
-  val targetLabel   = new Label("Target Aperture")
-  val targetUnits   = new Label("arcsec")
-  val target        = new NumberEdit(targetLabel, targetUnits, 2) <| {_.peer.setEnabled(false)}
-  new ButtonGroup(autoAperture, userAperture)
-
-  layout(new Label("Analysis Method:")) = new Constraints { gridx = 0; gridy = 0; anchor = Anchor.West; gridwidth = 5; insets = new Insets(0, 0, 5, 0) }
-  layout(targetLabel)                   = new Constraints { gridx = 0; gridy = 1; anchor = Anchor.West; insets = new Insets(0, 0, 0, 10) }
-  layout(autoAperture)                  = new Constraints { gridx = 1; gridy = 1; insets = new Insets(0, 0, 0, 3) }
-  layout(userAperture)                  = new Constraints { gridx = 2; gridy = 1; insets = new Insets(0, 0, 0, 5) }
-  layout(target)                        = new Constraints { gridx = 3; gridy = 1; anchor = Anchor.West;  insets = new Insets(0, 0, 0, 3) }
-  layout(targetUnits)                   = new Constraints { gridx = 4; gridy = 1; anchor = Anchor.West }
-  layout(skyLabel)                      = new Constraints { gridx = 0; gridy = 2; anchor = Anchor.West; insets = new Insets(0, 0, 0, 10) }
-  layout(sky)                           = new Constraints { gridx = 3; gridy = 2; anchor = Anchor.West; insets = new Insets(0, 0, 0, 3) }
-  layout(skyUnits)                      = new Constraints { gridx = 4; gridy = 2; anchor = Anchor.West }
-
-  listenTo(autoAperture, userAperture, target, sky)
-  reactions += {
-    case ButtonClicked(`autoAperture`)  => deafTo(userAperture); toggleUserAperture(enabled = false, "");   listenTo(userAperture); updateCache(); publish(new SelectionChanged(this))
-    case ButtonClicked(`userAperture`)  => deafTo(userAperture); toggleUserAperture(enabled = true, "2.0"); listenTo(userAperture); updateCache(); publish(new SelectionChanged(this))
-    case ValueChanged(_)                => updateCache(); publish(new SelectionChanged(this))
-  }
-
-  // get current nodeKey (used as lookup key for the observation)
-  def nodeKey = owner.getContextObservation.getNodeKey
-
-  // udpate the value that's currently cached for this observation
-  def updateCache() = analysisMethod.foreach { AnalysisMethodPanel.analysisCache.put(nodeKey, _) }
-
-  // generate a default analysis method for observations for which we don't have anything in the cache yet
-  def defaultMethod: AnalysisMethod = owner.getContextInstrumentDataObject match {
-      case i: InstGmosNorth => i.getFPUnitNorth match {
-        case FPUnitNorth.IFU_1                      => AutoAperture(500)  // GMOS-N IFU-2
-        case FPUnitNorth.IFU_2 | FPUnitNorth.IFU_3  => AutoAperture(250)  // GMOS-N IFU-B / IFU-R
-        case _                                      => AutoAperture(1)    // GMOS-N everything else
-      }
-      case i: InstGmosSouth => i.getFPUnitSouth match {
-        case FPUnitSouth.IFU_1                      => AutoAperture(500)  // GMOS-S IFU-2
-        case FPUnitSouth.IFU_2 | FPUnitSouth.IFU_3  => AutoAperture(250)  // GMOS-S IFU-B / IFU-R
-        case _                                      => AutoAperture(1)    // GMOS-S everything else
-      }
-      case i: InstNIRI                              => AutoAperture(1.0)
-      case i: Flamingos2                            => AutoAperture(1.0)
-      case i: Gsaoi                                 => AutoAperture(1.0)
-      case _                                        => defaultAnalysisMethod
-    }
-
-  // helper to activate/deactivate the user value for the target aperture
-  def toggleUserAperture(enabled: Boolean, v: String) = {
-    target.peer.setEnabled(enabled)
-    targetUnits.enabled = enabled
-    target.peer.setValue(v)
-  }
-
-  // update UI elements and values that are displayed
-  def update() = {
-    AnalysisMethodPanel.analysisCache.getOrElse(nodeKey, defaultMethod) match {
-      case AutoAperture(s) =>
-        autoAperture.selected = true
-        toggleUserAperture(enabled = false, "")
-        sky.peer.setValue(s)
-      case UserAperture(d, s) =>
-        userAperture.selected = true
-        toggleUserAperture(enabled = true, d.toString)
-        sky.peer.setValue(s)
-    }
-
-    // special handling for NIRI, F2 and GSAOI (OCSADV-345)
-    Option(owner.getContextInstrumentDataObject).foreach { _.getType match {
-      case InstNIRI.SP_TYPE | Flamingos2.SP_TYPE | Gsaoi.SP_TYPE  =>
-        sky.tooltip       = "This instrument does not support user defined values for the sky aperture."
-        sky.enabled       = false
-      case _  =>
-        sky.tooltip       = null
-        sky.enabled       = true
-    }}
-  }
-
-  // create the analysis method for the current user entries
-  def analysisMethod: Option[AnalysisMethod] =
-    if (autoAperture.selected) autoApertureValue else userApertureValue
-
-  private def autoApertureValue: Option[AnalysisMethod] =
-    sky.value.map(AutoAperture)
-
-  private def userApertureValue: Option[AnalysisMethod]  =
-    for {
-      sky <- sky.value
-      trg <- target.value
-    } yield UserAperture(trg, sky)
-
-}
-
-private class PlotDetailsPanel extends GridBagPanel {
-
-  val autoLimits      = new RadioButton("Auto") { focusable = false; background = Color.WHITE; selected = true }
-  val userLimits      = new RadioButton("User") { focusable = false; background = Color.WHITE }
-  val lowLimitLabel   = new Label("Low")
-  val lowLimitUnits   = new Label("nm")
-  val lowLimit        = new NumberEdit(lowLimitLabel, lowLimitUnits, 0)       { enabled = false }
-  val highLimitLabel  = new Label("High")
-  val highLimitUnits  = new Label("nm")
-  val highLimit       = new NumberEdit(highLimitLabel, highLimitUnits, 2000)  { enabled = false }
-  new ButtonGroup(autoLimits, userLimits)
-
-  background = Color.WHITE
-  layout(new Label("Limits:"))  = new Constraints { gridx = 0; gridy = 0; insets = new Insets(0, 0, 0, 20) }
-  layout(autoLimits)            = new Constraints { gridx = 1; gridy = 0; insets = new Insets(0, 0, 0, 10) }
-  layout(userLimits)            = new Constraints { gridx = 2; gridy = 0; insets = new Insets(0, 0, 0, 20) }
-  layout(lowLimitLabel)         = new Constraints { gridx = 3; gridy = 0; insets = new Insets(0, 0, 0, 10) }
-  layout(lowLimit)              = new Constraints { gridx = 4; gridy = 0; fill = Fill.Horizontal }
-  layout(lowLimitUnits)         = new Constraints { gridx = 5; gridy = 0; insets = new Insets(0, 5, 0, 0) }
-  layout(highLimitLabel)        = new Constraints { gridx = 6; gridy = 0; insets = new Insets(0, 20, 0, 10) }
-  layout(highLimit)             = new Constraints { gridx = 7; gridy = 0; fill = Fill.Horizontal }
-  layout(highLimitUnits)        = new Constraints { gridx = 8; gridy = 0; insets = new Insets(0, 5, 0, 0) }
-
-  listenTo(autoLimits, userLimits, lowLimit, highLimit)
-  reactions += {
-    case ButtonClicked(`autoLimits`)  => lowLimit.enabled = false; highLimit.enabled = false; publish(new SelectionChanged(this))
-    case ButtonClicked(`userLimits`)  => lowLimit.enabled = true;  highLimit.enabled = true;  publish(new SelectionChanged(this))
-    case ValueChanged(_)              => publish(new SelectionChanged(this))
-  }
-
-  def plottingDetails: PlottingDetails =
-    if (autoLimits.selected) PlottingDetails.Auto
-    else userPlottingDetails.getOrElse(PlottingDetails.Auto)
-
-  private def userPlottingDetails: Option[PlottingDetails] =
-    for {
-      low    <- lowLimit.value
-      high   <- highLimit.value
-      (l, h) <- if (low < high) Some((low, high)) else None
-    } yield new PlottingDetails(PlotLimits.USER, l, h)
-
-}
-
-// light weight wrapper to turn NumberBoxWidget into a Scala swing component
-private class NumberEdit(label: Label, units: Label, default: Double = 0) extends Component {
-  override lazy val peer = new NumberBoxWidget {
-    setColumns(6)
-    setValue(default)
-    setMinimumSize(getPreferredSize)
-    addWatcher(new TextBoxWidgetWatcher {
-      override def textBoxKeyPress(tbwe: TextBoxWidget): Unit = textBoxAction(tbwe)
-      override def textBoxAction(tbwe: TextBoxWidget): Unit =
-        try {
-          publish(new ValueChanged(NumberEdit.this))
-          tbwe.requestFocus()
-        } catch {
-          case _: NumberFormatException =>
-        }
-    })
-  }
-
-  override def enabled_=(e: Boolean) = {
-    label.enabled = e
-    peer.setEnabled(e)
-    units.enabled = e
-  }
-
-  def value: Option[Double] =
-    try {
-      Some(peer.getValue.toDouble)
-    } catch {
-      case _: NumberFormatException => None
-    }
-}
-
